% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnei_brak_use_case.R
\name{bnei_brak_use_case}
\alias{bnei_brak_use_case}
\title{Bnei Brak data}
\usage{
bnei_brak_use_case(
  initialization = 3,
  max_iter = 10,
  tolerance = 0.01,
  full_log = FALSE
)
}
\description{
estimation of the susceptibilities of adults/adolescents/children using a dataset containing SARS-CoV-2 testing results, collected from the city of Bnei Brak, Israel.
}
