% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_dict.R
\name{build_dict}
\alias{build_dict}
\title{Build household model from obs}
\usage{
build_dict(obs_list, risk_groups, past_dict = NULL)
}
\arguments{
\item{obs_list}{list of observations}

\item{risk_groups}{number of risk groups in the model}

\item{past_dict}{optional}
}
\description{
shell function for calculating the maximal household
and recursively calculate build the parametric model based on the given observation.
}
\examples{
obs_list <- list("1,2,0,1,1,0,1,1,0", "2,1,0,2,0,0,2,0,0")
build_dict(obs_list,3)

}
