% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_imputation_possibilities.R
\name{create_imputation_possibilities}
\alias{create_imputation_possibilities}
\title{Imputation table}
\usage{
create_imputation_possibilities(dict, id3_template, risk_groups)
}
\arguments{
\item{dict}{probabilistic model}

\item{id3_template}{partial observation}

\item{risk_groups}{number of risk groups in the model}
}
\description{
create all possible households for a partial observation and calculate,
based on the probabilistic model, the conditional probabilities
}
\examples{
obs_list <- list("1,2,0,1,1,0,1,1,0", "2,1,0,2,0,0,2,0,0")
create_imputation_possibilities(build_dict(obs_list,3), obs_list_to_dict(obs_list,3)[2,], 3)

}
