% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_imputations_table.R
\name{create_imputations_table}
\alias{create_imputations_table}
\title{Imputations table}
\usage{
create_imputations_table(dict, observation_table, risk_groups)
}
\arguments{
\item{dict}{probabilistic model}

\item{observation_table}{partial observations table}

\item{risk_groups}{number of risk groups in the model}
}
\description{
create all possible households for a table of partial observations and calculate, based on the probabilistic model, the conditional probabilities
}
\examples{
obs_list <- list("1,2,0,1,1,0,1,1,0", "2,1,0,2,0,0,2,0,0")
create_imputations_table(build_dict(obs_list,3), obs_list_to_dict(obs_list,3), 3)

}
