% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_symbolic_dictionary.R
\name{create_symbolic_dictionary}
\alias{create_symbolic_dictionary}
\title{Build household model}
\usage{
create_symbolic_dictionary(max_members, past_dict = NULL)
}
\arguments{
\item{max_members}{maximal members per risk group}

\item{past_dict}{optional}
}
\description{
recursively build the parametric model up to the maximal household.
}
