% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectation_phase.R
\name{expectation_phase}
\alias{expectation_phase}
\title{Expectation phase}
\usage{
expectation_phase(idsTable, Q, dig = 7)
}
\arguments{
\item{idsTable}{table of possible households and their parametric probabilities}

\item{Q}{previous estimator, to be evaluated by}

\item{dig}{precision of the returned values}
}
\description{
implement the expectation phase by evaluating the symbolic probabilities at a given vector Q
}
\examples{
obs_list <- list("1,2,0,1,1,0,1,1,0", "2,1,0,2,0,0,2,0,0")
dict <- build_dict(obs_list,3)
imputationsTable <- create_imputations_table(dict, obs_list_to_dict(obs_list,3), 3)
expectation_phase(imputationsTable, c(0.7, 0.8, 0.8), dig=4)

}
