% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_function_robust.R
\name{make_function_robust}
\alias{make_function_robust}
\title{Make function from string}
\usage{
make_function_robust(body, param_names = paste0("Q", 1:3))
}
\arguments{
\item{body}{string containing parametric math expression}

\item{param_names}{parameters of the UDF}
}
\description{
generate UDF from a string containing math expression
}
