% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparing_for_EM.R
\name{preparing_for_EM}
\alias{preparing_for_EM}
\title{Preparing for EM}
\usage{
preparing_for_EM(dict, observation_table, risk_groups)
}
\arguments{
\item{dict}{probabilistic model}

\item{observation_table}{partial observations table}

\item{risk_groups}{number of risk groups in the model}
}
\description{
create all possible households, and calculate both the conditional probabilities and the unconditional one
}
\examples{
obs_list <- list("1,2,0,1,1,0,1,1,0", "2,1,0,2,0,0,2,0,0")
preparing_for_EM(build_dict(obs_list,3), obs_list_to_dict(obs_list,3), 3)

}
