% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_Bnei_Brak.R
\name{preprocessing_Bnei_Brak}
\alias{preprocessing_Bnei_Brak}
\title{Preprocessing Bnei Brak data}
\usage{
preprocessing_Bnei_Brak(threshold = 2)
}
\arguments{
\item{threshold}{an individual result considered as certain if either
\enumerate{
\item a serological test result is available
\item at least one PCR positive result or
\item threshold negative results
}}
}
\description{
create all possible households, and calculate both conditional/unconditional probabilities
}
