% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_permutations.R
\name{prob_permutations}
\alias{prob_permutations}
\title{Prob permutations}
\usage{
prob_permutations(risk_groups, sym_prob, id2)
}
\arguments{
\item{risk_groups}{number of risk groups in the model}

\item{sym_prob}{symbolic probability of the original household}

\item{id2}{id2 of the original household}
}
\description{
calculating the symbolic probabilities of all permutations of the risk groups
}
\examples{
prob_permutations(3,"Q1*Q2","1,1,0,0,0,0")
prob_permutations(3, 'Q1^2*Q2', '2,1,0,0,0,0')

}
