% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_EM.R
\name{run_EM}
\alias{run_EM}
\title{Run EM}
\usage{
run_EM(
  imputationsTable,
  risk_groups,
  initialization = 3,
  max_iter = 10,
  tolerance = 0.01,
  full_log = FALSE
)
}
\arguments{
\item{imputationsTable}{weighted obs}

\item{risk_groups}{number of risk groups in the model}

\item{initialization}{number of starting points to consider}

\item{max_iter}{maximal steps to take from each starting point}

\item{tolerance}{stopping criterion}

\item{full_log}{whether to return the result from each step or just the best step}
}
\description{
run EM algorithm on the weighted observations to find the estimators
}
\examples{
obs_list <- list("1,2,0,1,1,0,1,1,0", "2,1,0,2,0,0,2,0,0", "0,0,1,0,0,0,0,0,0")
risk_groups <- 3
dict <- build_dict(obs_list,risk_groups)
number_of_families <- 15
new_obs_list <- generate_obs_list(dict, risk_groups, number_of_families, families_generator = c(2,1,1), models_generator = c(0.8, 0.8, 0.87), missing_mechanism = 0.7, seed = 3)
observations_dict <- obs_list_to_dict(new_obs_list, risk_groups)
imputationsTable <- preparing_for_EM(dict, observations_dict, risk_groups)
run_EM(imputationsTable, risk_groups)

run_EM(imputationsTable, risk_groups, full_log=TRUE)

}
