% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_max_members.R
\name{calc_max_members}
\alias{calc_max_members}
\title{Calculate maximal household}
\usage{
calc_max_members(obs_list, risk_groups)
}
\arguments{
\item{obs_list}{list of observations}

\item{risk_groups}{number of risk groups in the model}
}
\description{
calculating the maximal household, element-wise, based on the given observation.
}
