% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_obs_list.R
\name{generate_obs_list}
\alias{generate_obs_list}
\title{Generate observations}
\usage{
generate_obs_list(
  model_dict,
  risk_groups,
  number_of_families,
  families_generator = c(2, 1, 1),
  models_generator = c(0.8, 0.8, 0.87),
  missing_mechanism = 0.8,
  seed = 3
)
}
\arguments{
\item{model_dict}{probabilistic model}

\item{risk_groups}{number of risk groups in the model}

\item{number_of_families}{number of observations to generate}

\item{families_generator}{parameters for the poisson distribution, for generating household members}

\item{models_generator}{parameters for the parametric model}

\item{missing_mechanism}{the generated observations will contain 1-missing_mechanism\% missing tests}

\item{seed}{seed is used for reproducibility}
}
\description{
generate household observations based on a given probabilistic model.
}
\examples{
obs_list <- list("1,2,0,1,1,0,1,1,0", "2,1,0,2,0,0,2,0,0")
generate_obs_list(build_dict(obs_list,3), 3, 4)
generate_obs_list(build_dict(obs_list,3), 3, 4, missing_mechanism = 1)

}
