% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs_list_to_dict.R
\name{obs_list_to_dict}
\alias{obs_list_to_dict}
\title{Observations list to dict}
\usage{
obs_list_to_dict(obs_list, risk_groups)
}
\arguments{
\item{obs_list}{list of observations}

\item{risk_groups}{number of risk groups in the model}
}
\description{
transfers obs_list to a dictionary.
}
