% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_Bnei_Brak.R
\name{preprocessing_Bnei_Brak}
\alias{preprocessing_Bnei_Brak}
\title{Preprocessing Bnei Brak data}
\usage{
preprocessing_Bnei_Brak(threshold = 2)
}
\arguments{
\item{threshold}{an individual result considered as certain if either 1. a serological test result is available 2. at least one PCR positive result or 3. threshold negative results}
}
\description{
create all possible households, and calculate both the conditional probabilities and the unconditional one
}
